refactor: HomePage UI 영역 컴포넌트화하여 구조 개선
- Footer를 common/Footer로 분리하여 재사용 가능하도록 구조화
- 상품 로딩 카드 UI를 ProductSkeletonCard로 분리
- 공통 로딩 인디케이터를 LoadingIndicator 컴포넌트로 분리
- 검색 및 필터 영역을 ProductFilterPanel로 분리
- HomePage에서 각 컴포넌트 호출 방식으로 변경하여 가독성과 유지보수성 향상
